Extract to cache_utils
