+++ app.rb
@@ -9,6 +9,7 @@ Bundler.require(:default, ENV["RACK_ENV"])
 
 require "rollbar/middleware/sinatra"
 require_relative "./lib/on_air_bot"
+require_relative "./lib/cache_utils"
 
 class App < Sinatra::Base
   use Rollbar::Middleware::Sinatra
@@ -32,7 +33,9 @@ class App < Sinatra::Base
     slim :index
   end
 
-  helpers do # rubocop:disable Metrics/BlockLength
+  helpers CacheUtils
+
+  helpers do
     def girl_birthdays(from_date, to_date)
       date_girls = {}
       girls = Precure.all.select(&:have_birthday?)
@@ -55,44 +58,6 @@ class App < Sinatra::Base
       today = Time.current.to_date
       fetch_cache("precure_programs") do
         OnAirBot.programs(today, today + PROGRAM_WEEKS.weeks).select { |program| program[:title].include?(OnAirBot::NOTIFY_TITLE) }
-      end
-    end
-
-    def fetch_cache(key)
-      cache = cache_client
-
-      begin
-        cached_response = cache.get(key)
-        return cached_response if cached_response
-      rescue => e
-        logger.warn(e)
-        Rollbar.warning(e)
-      end
-
-      response = yield
-
-      begin
-        cache.set(key, response)
-      rescue => e
-        logger.warn(e)
-        Rollbar.warning(e)
-      end
-
-      response
-    end
-
-    def cache_client
-      options = { namespace: Global.memcached.namespace, compress: true, expires_in: Global.memcached.expire_minutes.minutes }
-
-      Dalli.logger.level = Logger::WARN
-
-      if ENV["MEMCACHEDCLOUD_SERVERS"]
-        # Heroku
-        options[:username] = ENV["MEMCACHEDCLOUD_USERNAME"]
-        options[:password] = ENV["MEMCACHEDCLOUD_PASSWORD"]
-        Dalli::Client.new(ENV["MEMCACHEDCLOUD_SERVERS"].split(","), options)
-      else
-        Dalli::Client.new("#{Global.memcached.host}:#{Global.memcached.port}", options)
       end
     end
   end
