+++ spec/spec_helper.rb
@@ -16,6 +16,11 @@ def spec_dir
   Pathname(__dir__)
 end
 
+Global.configure do |config|
+  config.environment = ENV["RACK_ENV"]
+  config.config_directory = "#{__dir__}/../config/global"
+end
+
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
 # The generated `.rspec` file contains `--require spec_helper` which will cause
@@ -122,5 +127,11 @@ RSpec.configure do |config|
 
   config.after do
     Timecop.return
+  end
+
+  include CacheUtils
+
+  config.before do
+    cache_client.flush_all
   end
 end
